spoc-robot-training/requirements.txt
# Checks style, syntax, and other useful errors.
flake8

# Static type checking
mypy==1.2.0

# Automatic code formatting
black

# Running tests
pytest
pytest-xdist

# Flaky tests
flaky

# Other reqs
invoke==2.0.0
attrs
prior
stringcase
nltk @ git+https://github.com/nltk/nltk@582e6e35f0e6c984b44ec49dcb8846d9c011d0a8
phonemizer
networkx
numpy-quaternion==2022.4.1
tensorboardx>=2.1
setproctitle
moviepy>=1.0.3
filelock
numpy==1.23.5
matplotlib>=3.3.1
opencv-python
wheel>=0.36.2
attrs>=21.4.0
scipy>=1.5.4
prior
canonicaljson
plotly
shapely
h5py
pyquaternion>=0.9.9
omegaconf
boto3
petname
wget
pandas
python-sat
python-fcl
invoke
wandb
scikit-video
nbformat
pre-commit
black==23.3.0
scikit-image
torchmetrics
av
shortuuid
transformers
ipdb
decorator==4.4.2
timeout-decorator
objathor>=0.0.2

# for objaverse pipeline
trimesh
objaverse
compress_pickle
compress_json

# for proper determiners with phonemization
phonemizer

allenact==0.5.4
allenact_plugins[all]
#torch==2.0.1
#torchvision
xformers
lightning
open-clip-torch
